+++ build-logic/gobley-gradle-cargo/src/main/kotlin/CargoPlugin.kt
@@ -214,6 +214,7 @@ class CargoPlugin : Plugin<Project> {
                 tasks.register<RustUpTargetAddTask>({ +cargoBuild.rustTarget }) {
                     group = TASK_GROUP
                     this.rustTarget.set(cargoBuild.rustTarget)
+                    this.rustVersion.set(cargoExtension.rustVersion)
                 }
             for (cargoBuildVariant in cargoBuild.variants) {
                 val projectLayout = layout