+++ build-logic/gobley-gradle/src/main/kotlin/rust/targets/RustAndroidTarget.kt
@@ -61,6 +61,9 @@ enum class RustAndroidTarget(
     override val friendlyName = "Android$name"
 
     override val supportedKotlinPlatformTypes = arrayOf(KotlinPlatformType.androidJvm)
+
+    override fun tier(rustVersion: String) = 2
+
     override fun outputFileName(crateName: String, crateType: CrateType): String? =
         crateType.outputFileNameForLinux(crateName)