+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased](https://github.com/gobley/gobley/compare/v0.1.0...HEAD)
 
+### New Features
+
+- Stub generation for unsupported JS/WASM targets ([#29](https://github.com/gobley/gobley/pull/29)).
+- Kotlin/Native tvOS & watchOS support ([#35](https://github.com/gobley/gobley/pull/35)).
+
 ## [0.1.0](https://github.com/gobley/gobley/releases/tag/v0.1.0) - 2025-03-03
 
 ### New Features
@@ -9,6 +14,9 @@
 - Support for `uniffi@0.28.3`.
 - Support for Kotlin 2.x.
 - Gradle plugins for cross-compilation automation.
+  - Android
+  - JVM
+  - Kotlin/Native (MinGW, macOS, Linux, & iOS)
 
 ## [UniFFI Kotlin Multiplatform Bindings 0.1.0](https://gitlab.com/trixnity/uniffi-kotlin-multiplatform-bindings/-/tags/v0.1.0) - 2023-11-25